Epic, Story, Feature annotation @Inherited does not work
Hi guys. In your code, there is a bug. If I use Epic, Story, or Feature annotation in my Tests and put these annotations in test abstract class, these annotations do not work at all. I analyzed the code and found out that it's because you get declared annotations and do not get annotations that are inherited